image_service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
requests
python-multipart